fix somthong

+++ api/edu/src/main/java/com/group4/edu/service/Impl/SudentServiceImpl.java
@@ -23,10 +23,7 @@ import javax.persistence.EntityManager;
 import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -296,18 +293,20 @@ public class SudentServiceImpl implements StudentService {
         return studentDtos;
     }
 
+    @Transactional
     @Override
     public boolean deleteStTT(Long id) {
         Student student = studentRepository.findById(id).orElse(null);
-        if(student.getStudentType().equals(EduConstants.StudentType.STUDENT_TT.getValue())){
+        if(student == null){
+            return false;
+        }
+        if (student.getStudentType().equals(EduConstants.StudentType.STUDENT_TT.getValue())) {
             Account account = accountRepository.getAccountByUserId(student.getId()).orElse(null);
             accountRepository.delete(account);
-
             List<Internship> internships = intershipRepository.getInternshipByStudentId(student.getId());
             intershipRepository.deleteAll(internships);
-
-            studentRepository.deleteById(student.getId());
-            System.out.println(student.getId());
+            if (studentRepository.existsById(student.getId()))
+                studentRepository.deleteById(student.getId());
             return true;
         }
         if(student.getStudentType().equals(EduConstants.StudentType.ALL.getValue())){
@@ -315,11 +314,7 @@ public class SudentServiceImpl implements StudentService {
             Account account = accountRepository.getAccountByUserId(student.getId()).orElse(null);
             if(account != null){
                 Set<Role> roleSet = account.getRoles();
-                for(Role role: roleSet){
-                    if(role.getCode().equals(EduConstants.Role.ROLESTUDENT_TT.getKey())){
-                        roleSet.remove(role);
-                    }
-                }
+                roleSet.removeIf(role -> role.getCode().equals(EduConstants.Role.ROLESTUDENT_TT.getKey()));
             }
             List<Internship> internships = intershipRepository.getInternshipByStudentId(student.getId());
             intershipRepository.deleteAll(internships);
